LPS-83561 - Moves code to management toolbar display context
